Allow RPC methods with imported input types
Re: https://github.com/fiorix/protoc-gen-cobra/issues/2

The input type is assigned in the handler using the local type name.
This works fine if the type is part of the same package, but throws a
name error if the type was imported.

This makes two changes. The first is to check the the package of input
types to see if it's the same as the generated file's package (ie local)
and prefixes the type name if appropriate.

The second is to add import statements for any input types. This
requires a change to the `Plugin` interface, as `GenerateImports` has no
knowledge of the file's dependencies. This adds a second argument to
that method and adds import statements for any input types from exported
packages.